requirements.txt
opencv-python
ultralytics
numpy
pyyaml